+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import Current from "./Current";
+import { killAllChildProcesses } from "./execShell";
 import { SwiftLint } from "./SwiftLintProvider";
 
 const swiftLint = new SwiftLint();
@@ -10,4 +11,8 @@ export function activate(context: vscode.ExtensionContext) {
   }
 
   swiftLint.activate(context);
+}
+
+export function deactivate(_context: vscode.ExtensionContext) {
+  killAllChildProcesses();
 }