+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Pass correct working directory to SwiftLint #24
+- Kill remaining SwiftLint processes on deactivate #24
 
 ## 1.2.6